refactor: Replace rule's config processing with config validation
The `validateConfig` is used to trigger errors when rule's config is not valid. Previously issues were reported for invalid config, which is bad for UX as those issues were lost in the middle of real lint issues. By throwing an error we can stop the linter and tell the user that something is not good.